Move OpaqueRepr into its own module.
